backend/summarizer_agent/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
pydantic==2.11.7
requests==2.31.0
python-dotenv==1.0.1